Crash fixed, midi note on/off pattern still doesn't do anything

+++ LedStrip.cpp
@@ -46,7 +46,7 @@ struct FastLedCRGB* LedStrip::GetLeds()
 }
 
 
-struct FastLedCRGB* LedStrip::GetLed(uint16_t ledIndex)
+struct FastLedCRGB* LedStrip::GetLed(uint8_t ledIndex)
 {
 #ifdef _DEBUG
 	if (ledIndex >= _nrOfLeds)
@@ -60,7 +60,7 @@ struct FastLedCRGB* LedStrip::GetLed(uint16_t ledIndex)
 
 void LedStrip::SetAllLeds(LedColor::EColor color, uint8_t step)
 {
-	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		LedColor::SetRgb(&led.red, &led.green, &led.blue, color, step);
@@ -70,7 +70,7 @@ void LedStrip::SetAllLeds(LedColor::EColor color, uint8_t step)
 
 void LedStrip::SetAllLeds(uint32_t color)
 {
-	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		LedColor::SetRgb(&led.red, &led.green, &led.blue, color);
@@ -80,7 +80,7 @@ void LedStrip::SetAllLeds(uint32_t color)
 
 void LedStrip::SetAllLeds(uint8_t red, uint8_t green, uint8_t blue)
 {
-	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		LedColor::SetRgb(&led.red, &led.green, &led.blue, red, green, blue);
@@ -109,7 +109,7 @@ void LedStrip::SetLed(struct FastLedCRGB* led, uint8_t red, uint8_t green, uint8
 uint16_t LedStrip::CalculateCurrentInMilliAmp()
 {
 	uint32_t brightness = 0;
-	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		brightness += led.red + led.green + led.blue;
